requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
oauthlib==3.2.0
python-dotenv==0.20.0
requests==2.28.1
requests-oauthlib==1.3.1
tweepy==4.10.0
urllib3==1.26.10
tensorflow==2.9.1
transformers==4.20.1
sentencepiece==0.1.96
Pillow==9.2.0